Refactored types and updated dep-to mctest-v0.5.4 v0.3.2

+++ cachetypes.go
@@ -8,7 +8,7 @@ type ValueType interface{}
 
 type CacheValue struct {
 	value  ValueType
-	expire uint
+	expire int64
 }
 
 type CacheResponse struct {
